Add load-project, and make distill print changes
The load-project function will read the current project.clj file's
dependencies and load any new ones.

The distill function now prints the dependency coordinates that are
loaded, and separately those that are not due to conflicts.  It also
now accepts a sequence of dependencies to load.

The previous return behaviour of distill is available in distill*.